Now 3 rays are shoot to check top and bottom intersection
